requirements.txt
absl-py==2.3.1
accelerate==1.1.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.49.1
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
distlib==0.4.0
einops==0.8.2
exceptiongroup==1.3.1
executing==2.2.1
facexlib==0.3.0
filelock==3.20.3
filterpy==1.4.5
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
icecream==2.1.10
identify==2.6.16
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
lmdb==1.7.5
Markdown==3.10.1
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.4.2
nodeenv==1.10.0
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-clip==1.0.1
opencv-python==4.13.0.90
opencv-python-headless==4.13.0.90
packaging==26.0
pandas==2.3.3
peft==0.9.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiqa==0.1.14.1
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
ruff==0.14.14
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
sentry-sdk==2.52.0
shapely==2.1.2
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
stringzilla==4.6.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.24
tokenizers==0.15.2
tomli==2.4.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.37.2
triton==3.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.36.1
wandb==0.25.0
wcwidth==0.5.0
Werkzeug==3.1.5
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
